+++ server/core_user.go
@@ -179,7 +179,7 @@ func DeleteUser(ctx context.Context, tx *sql.Tx, userID uuid.UUID) (int64, error
 	return res.RowsAffected()
 }
 
-func BanUsers(ctx context.Context, logger *zap.Logger, db *sql.DB, sessionCache SessionCache, ids []uuid.UUID) error {
+func BanUsers(ctx context.Context, logger *zap.Logger, db *sql.DB, config Config, sessionCache SessionCache, sessionRegistry SessionRegistry, tracker Tracker, ids []uuid.UUID) error {
 	statements := make([]string, 0, len(ids))
 	params := make([]interface{}, 0, len(ids))
 	for i, id := range ids {
@@ -195,6 +195,15 @@ func BanUsers(ctx context.Context, logger *zap.Logger, db *sql.DB, sessionCache 
 	}
 
 	sessionCache.Ban(ids)
+
+	for _, id := range ids {
+		// Disconnect.
+		for _, presence := range tracker.ListPresenceIDByStream(PresenceStream{Mode: StreamModeNotifications, Subject: id}) {
+			if err = sessionRegistry.Disconnect(ctx, presence.SessionID, true); err != nil {
+				return err
+			}
+		}
+	}
 
 	return nil
 }