+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 ## [Unreleased]
 ### Added
 - Add tournament record delete runtime functions and API.
+- Add insecure flag to runtime http functions to optionally skip TLS checks.
+- Add Satori API integration to all runtimes.
 
 ### Changed
 - Improve graceful shutdown of Google IAP receipt processor.
@@ -17,6 +19,8 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 - Consistent validation of override operator in runtime leaderboard record writes.
 - Correctly filter open/closed groups in the listing API.
 - Ensure direct message channel message listing is correctly scoped to participants only.
+- Make next and previous cursor of leaderboard and tournament records around owner operations consistent with record listing.
+- Make next and previous cursor of leaderboard and tournament records haystack operations consistent with record listing.
 
 ## [3.15.0] - 2023-01-04
 ### Added
@@ -74,6 +78,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 - Token and credentials as inputs on unlink operations are now optional.
 - Improve runtime IAP operation errors to include provider payload in error message.
 - Build with Go 1.19.2 release.
+- Disconnect users when they are banned from the console or runtime functions.
 
 ### Fixed
 - Observe the error if returned in storage list errors in JavaScript runtime.